Move local video to its own file
