+++ src/peerConnection.ts
@@ -13,7 +13,7 @@ import { updateBandwidthRestriction } from './util/sdpInject'
 import { getIceServers } from './iceServerData'
 import { onResolutionChange } from './selectDevice'
 import { updateAllParameters, updateParameters, updateResolution } from './senderParameters'
-import { localVideo } from './custom-elements/floating-video'
+import { localVideo } from './localVideo'
 import { closeShareDialog, isShareDialogOpen } from './shareDialog'
 import { stream } from './peerConnectionsManager'
 import { peerConnections } from './peerConnectionsManager'
